requirements.txt
hstrader
pandas
plotly
python-dotenv
dash